requirements.txt
streamlit==1.32.0
fastapi==0.109.0
uvicorn==0.27.0
keras==3.0.5
tensorflow==2.15.0
numpy==1.26.3
Pillow==10.2.0
python-multipart==0.0.6
requests==2.31.0